refactor: align role-a admin routes and pages
